BIG UPDATE!!
this time i am way to lazy to list up all of the changes so feel free to explore by yourself

thanks Hideshichan for reporting bugs

+++ src/main/kotlin/noammaddons/config/CustomMainMenu/TextButton.kt
@@ -7,6 +7,7 @@ import noammaddons.utils.RenderUtils.drawRoundedBorder
 import noammaddons.utils.RenderUtils.drawRoundedRect
 import java.awt.Color
 
+
 class TextButton(val text: String, var x: Int, var y: Int) {
 	val w: Int = 150
 	val h: Int = 20
@@ -19,27 +20,23 @@ class TextButton(val text: String, var x: Int, var y: Int) {
 		if (isHovered(mouseX, mouseY)) {
 			drawRoundedBorder(
 				Color(255, 255, 255),
-				x.toDouble(),
-				y.toDouble(),
-				w.toDouble(),
-				h.toDouble(),
-				5.0, 1.0, 2
+				x.toFloat(), y.toFloat(),
+				w.toFloat(), h.toFloat(),
+				5f, 1f, 2f
 			)
 		}
 		
 		drawRoundedRect(
 			Color(33, 33, 33, 255),
-			x.toDouble(),
-			y.toDouble(),
-			w.toDouble(),
-			h.toDouble()
+			x.toFloat(), y.toFloat(),
+			w.toFloat(), h.toFloat()
 		)
 		
 		drawCenteredText(
 			text,
-			(x + w / 2f).toDouble(),
-			(y + h / 2f - 4).toDouble(),
-			1.0, Color(0, 255, 255)
+			x + w / 2f,
+			y + h / 2f - 4,
+			1f, Color(0, 255, 255)
 		)
 		
 		GlStateManager.color(1.0f, 1.0f, 1.0f, 1f)
@@ -51,8 +48,7 @@ class TextButton(val text: String, var x: Int, var y: Int) {
 			mouseY.toFloat(),
 			x.toDouble(),
 			y.toDouble(),
-			w,
-			h
+			w, h
 		)
 	}
 }
